requirements.txt
streamlit
beautifulsoup4
requests
duckduckgo-search
openai
fake_useragent
python-dotenv
